+++ taf/tools/repo/__init__.py
@@ -191,14 +191,10 @@ def attach_to_group(group):
     @click.option("--expected-repo-type", default="either", type=click.Choice(["test", "official", "either"]),
                   help="Indicates expected authentication repository type - test or official. If type is set to either, "
                   "the updater will not check the repository's type")
-    @click.option("--error-if-unauthenticated-repos-list", multiple=True, help="Raise an error if the repository allows "
-                  "unauthentiated commits and the updater detected authenticated commits newer than local "
-                  "head commit")
     @click.option("--scripts-root-dir", default=None, help="Scripts root directory, which can be used to move scripts "
                   "out of the authentication repository for testing purposes (avoid dirty index). Scripts will be expected "
                   "to be located in scripts_root_dir/repo_name directory")
-    def update(url, clients_auth_path, clients_library_dir, default_branch, from_fs, expected_repo_type, error_if_unauthenticated_repos_list,
-               scripts_root_dir):
+    def update(url, clients_auth_path, clients_library_dir, default_branch, from_fs, expected_repo_type, scripts_root_dir):
         """
         Update and validate local authentication repository and target repositories. Remote
         authentication's repository url needs to be specified when calling this command. If the
@@ -219,11 +215,6 @@ def attach_to_group(group):
         if this flag is omitted in the mentioned case. Do not use this flag when validating a non-test
         repository as that will also result in an error.
 
-        Some repositories can contain unauthenticated commits in-between two authenticated ones. The updater
-        will check existence and order of all authenticated commits and ignore unauthenticated. To raise
-        an error if there are new unauthenticated commits (newer than the last local commit),
-        error-if-unauthenticated option can be used.
-
         Scripts root directory option can be used to move scripts out of the authentication repository for
         testing purposes (avoid dirty index). Scripts will be expected  o be located in scripts_root_dir/repo_name directory
         """
@@ -231,8 +222,7 @@ def attach_to_group(group):
             raise click.UsageError('Must specify either authentication repository path or library directory!')
 
         update_repository(url, clients_auth_path, clients_library_dir, default_branch, from_fs,
-                          UpdateType(expected_repo_type),
-                          error_if_unauthenticated_repos_list=error_if_unauthenticated_repos_list, scripts_root_dir=scripts_root_dir)
+                          UpdateType(expected_repo_type), scripts_root_dir=scripts_root_dir)
 
     @repo.command()
     @click.argument("clients-auth-path")