+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Added
 
+### Changed
+
+### Fixed
+
+## [0.16.0] - 04/16/2022
+
+### Added
+
 - Add `allow_unsafe` flag to git repo class as a response to a recent git security fix ([229])
 ### Changed
 
@@ -614,7 +622,8 @@ and this project adheres to [Semantic Versioning][semver].
 [keepachangelog]: https://keepachangelog.com/en/1.0.0/
 [semver]: https://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/openlawlibrary/taf/compare/v0.15.0...HEAD
+[Unreleased]: https://github.com/openlawlibrary/taf/compare/v0.16.0...HEAD
+[0.14.0]: https://github.com/openlawlibrary/taf/compare/v0.15.0...v0.16.0
 [0.14.0]: https://github.com/openlawlibrary/taf/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/openlawlibrary/taf/compare/v0.13.4...v0.14.0
 [0.13.4]: https://github.com/openlawlibrary/taf/compare/v0.13.3...v0.13.4